refactor and fs and demo and hax
