+++ CSC Assistant.Algo/CSC Assistant.Algo.fsproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>CSC_Assistant.Algo</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="Library.fs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CSC Assistant.Common\CSC Assistant.Common.csproj" />
+  </ItemGroup>
+
+</Project>
