+++ CSC Assistant.Common/CSC Assistant.Common.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>CSC_Assistant.Common</RootNamespace>
+  </PropertyGroup>
+
+</Project>
