+++ CSC Assistant.Common/DataStructures/Blob.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace CSC_Assistant.Client.DataStructures
+namespace CSC_Assistant.Common.DataStructures
 {
     public class BlobForDisplay
     {
