+++ CSC Assistant.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSC Assistant.Client", "CSC
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSC Assistant.Calculations", "CSC Assistant.Calculations\CSC Assistant.Calculations.csproj", "{3AD8A6F2-C320-4807-B6A7-CE16CEB93C70}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSC Assistant.Common", "CSC Assistant.Common\CSC Assistant.Common.csproj", "{26183DEF-C879-415C-859C-084E08C05447}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "CSC Assistant.Algo", "CSC Assistant.Algo\CSC Assistant.Algo.fsproj", "{3B0C9F8B-278F-490D-8700-AA6037557340}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{3AD8A6F2-C320-4807-B6A7-CE16CEB93C70}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3AD8A6F2-C320-4807-B6A7-CE16CEB93C70}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3AD8A6F2-C320-4807-B6A7-CE16CEB93C70}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26183DEF-C879-415C-859C-084E08C05447}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26183DEF-C879-415C-859C-084E08C05447}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26183DEF-C879-415C-859C-084E08C05447}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26183DEF-C879-415C-859C-084E08C05447}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B0C9F8B-278F-490D-8700-AA6037557340}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B0C9F8B-278F-490D-8700-AA6037557340}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B0C9F8B-278F-490D-8700-AA6037557340}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B0C9F8B-278F-490D-8700-AA6037557340}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
